Move @MarketAgnostic to rest.conf subpackage
