+++ server/src/main/java/com/xinra/reviewcommunity/rest/AuthController.java
@@ -1,5 +1,8 @@
 package com.xinra.reviewcommunity.rest;
 
+import com.xinra.reviewcommunity.auth.AccessRequires;
+import com.xinra.reviewcommunity.auth.Role;
+import com.xinra.reviewcommunity.rest.conf.MarketAgnostic;
 import javax.servlet.http.HttpServletRequest;
 import org.springframework.security.web.csrf.CsrfToken;
 import org.springframework.web.bind.annotation.RequestMapping;
